requirement.txt
python-dotenv
keyring
tastytrade
matplotlib